+++ cypress/locators/drawer/locators.js
@@ -1,0 +1,3 @@
+// component preview locators
+export const DRAWER = '[data-element="drawer"]';
+export const DRAWER_TOGGLE = '[data-element="drawer-toggle"]';