+++ src/components/accordion/accordion.spec.js
@@ -25,7 +25,8 @@ const isExpanded = (wrapper) => {
   }, wrapper.find(StyledAccordionIcon));
   assertStyleMatch({
     visibility: undefined,
-    maxHeight: `${contentHeight}px`
+    maxHeight: `${contentHeight}px`,
+    height: `${contentHeight}px`
   }, wrapper.find(StyledAccordionContentContainer));
 };
 
@@ -35,7 +36,8 @@ const isCollapsed = (wrapper) => {
   }, wrapper.find(StyledAccordionIcon));
   assertStyleMatch({
     visibility: 'hidden',
-    maxHeight: '0px'
+    maxHeight: '0px',
+    height: '0px'
   }, wrapper.find(StyledAccordionContentContainer));
 };
 
@@ -82,6 +84,22 @@ describe('Accordion', () => {
       const ev = { which: keyCode };
       wrapper.find(StyledAccordionTitleContainer).prop('onKeyDown')(ev);
       expect(onChange).toHaveBeenCalledWith(ev, true);
+    });
+
+    it('adjusts accordions height when the content changes', () => {
+      act(() => render({ expanded: true }));
+      wrapper.update();
+      isExpanded(wrapper);
+
+      const newContentHeight = 400;
+      jest.spyOn(
+        wrapper.find(StyledAccordionContent).getDOMNode(), 'scrollHeight', 'get'
+      ).mockImplementation(() => newContentHeight);
+      wrapper.setProps({ children: <div /> });
+      wrapper.update();
+      assertStyleMatch({
+        maxHeight: `${newContentHeight}px`
+      }, wrapper.find(StyledAccordionContentContainer));
     });
   });